Move signRequestStore tests to own file
