+++ src/test/kotlin/nl/nuts/discovery/service/LocalCertificateAndKeyServiceTest.kt
@@ -19,10 +19,6 @@
 
 package nl.nuts.discovery.service
 
-import net.corda.core.identity.CordaX500Name
-import nl.nuts.discovery.TestUtils
-import nl.nuts.discovery.store.SignRequestStore
-import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.springframework.beans.factory.annotation.Autowired
@@ -30,25 +26,14 @@ import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.junit4.SpringRunner
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
-import kotlin.test.assertNull
 
 
 @RunWith(SpringRunner::class)
 @SpringBootTest
 class LocalCertificateAndKeyServiceTest {
-    @Autowired
-    lateinit var service: CertificateAndKeyService
 
     @Autowired
     lateinit var certService: CertificateAndKeyService
-
-    @Autowired
-    lateinit var signRequestStore: SignRequestStore
-
-    @Before
-    fun setup() {
-
-    }
 
     @Test
     fun `the intermediate certificate is loaded from test resources`() {
@@ -66,44 +51,8 @@ class LocalCertificateAndKeyServiceTest {
     }
 
     @Test
-    fun `a request for signature is not automatically signed`() {
-        val subject = CordaX500Name.parse("O=Org,L=Gr,C=NL")
-        val req = TestUtils.createCertificateRequest(subject)
-
-        signRequestStore.addSigningRequest(req)
-        val certificate = signRequestStore.signedCertificate(subject)
-        val pendingCertificate = signRequestStore.pendingSignRequest(subject)
-
-        assertNull(certificate)
-        assertNotNull(pendingCertificate)
-    }
-
-    @Test
-    fun `a pending requests request can be retrieved`() {
-        val subject = CordaX500Name.parse("O=Org,L=Gr,C=NL")
-        val req = TestUtils.createCertificateRequest(subject)
-
-        signRequestStore.addSigningRequest(req)
-        val pendingRequests = signRequestStore.pendingSignRequests()
-        assertEquals(pendingRequests.size, 1)
-        assertEquals(subject, pendingRequests[0].legalName())
-    }
-
-    @Test
-    fun `a signed signature can be retrieved`() {
-        val subject = CordaX500Name.parse("O=Org,L=Gr,C=NL")
-        val req = TestUtils.createCertificateRequest(subject)
-        val signRequest = signRequestStore.addSigningRequest(req)
-        service.signCertificate(req)
-        signRequestStore.markAsSigned(signRequest)
-
-        val certs = signRequestStore.signedCertificates()
-        assertEquals(certs.size, 1)
-        assertEquals(subject, certs[0].legalName())
-    }
-
-    @Test
     fun `service validates with existing certificates and keys`() {
-        assertEquals(0, service.validate().size)
+        assertEquals(0, certService.validate().size)
     }
+
 }
